# src/models/loginDto.ts
export default class LoginDto{

    public userId?: string;
    public username?: string;
    public role?: string;
    public password?: string;
    public jwt?: string;

    constructor() {

    }
    
}